requirements.txt
Flask==2.0.1
textblob==0.15.3
googletrans==4.0.0-rc1
pyspellchecker==0.6.2
twilio==6.66.0
deep_translator==1.5.0
facebook-sdk==3.1.0
pyenchant==3.2.1
openai==0.27.0
